Moved files, Added Stones to obstacles, Set their damage
