+++ Assets/Prefabs/Obstacles/Stone1.prefab
@@ -13,9 +13,10 @@ GameObject:
   - component: {fileID: 3789719880210897352}
   - component: {fileID: 3931924094891786685}
   - component: {fileID: 773584195282589071}
+  - component: {fileID: -7972003275434162904}
   m_Layer: 7
   m_Name: Stone1
-  m_TagString: Untagged
+  m_TagString: Obstacle
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
@@ -199,3 +200,18 @@ PolygonCollider2D:
       - {x: 0.746858, y: -0.41161305}
       - {x: 0.93116117, y: -0.2441807}
   m_UseDelaunayMesh: 0
+--- !u!114 &-7972003275434162904
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 5646802508109994502}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: e07d0218e9b7bba449fcde74ba69656f, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  initialSpeed: 3
+  damage: 5
+  currentSpeed: 0
