--- Assets/Scenes/BrayScene.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: dc40d861cad1edf40923acb41096fc3d
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
