+++ Assets/Scripts/Spawnables/StoneObstacle.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e07d0218e9b7bba449fcde74ba69656f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
